code cleanup, remove deprecated lodash.get, file renames for clarity
